The model catalog and setup routes live in app/api/models.py; the fakes still bite, proven by a canary
